Moving PopoverComponent to Beta::Popover
