+++ app/components/primer/primer.pcss
@@ -16,11 +16,11 @@
 @import "./beta/label.pcss";
 @import "./beta/link.pcss";
 @import "./beta/blankslate.pcss";
+@import "./beta/popover.pcss";
 @import "./beta/progress_bar.pcss";
 @import "./beta/truncate.pcss";
 @import "./dropdown.pcss";
 @import "./menu_component.pcss";
-@import "./popover_component.pcss";
 @import "./state_component.pcss";
 @import "./subhead_component.pcss";
 @import "./timeline_item_component.pcss";
