+++ static/audited_at.json
@@ -50,6 +50,7 @@
   "Primer::Beta::IconButton": "",
   "Primer::Beta::Label": "",
   "Primer::Beta::Link": "",
+  "Primer::Beta::Popover": "",
   "Primer::Beta::ProgressBar": "",
   "Primer::Beta::Text": "",
   "Primer::Beta::Truncate": "",
