+++ static/statuses.json
@@ -50,6 +50,7 @@
   "Primer::Beta::IconButton": "beta",
   "Primer::Beta::Label": "beta",
   "Primer::Beta::Link": "beta",
+  "Primer::Beta::Popover": "beta",
   "Primer::Beta::ProgressBar": "beta",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
@@ -77,7 +78,7 @@
   "Primer::Navigation::TabComponent": "alpha",
   "Primer::OcticonComponent": "beta",
   "Primer::OcticonSymbolsComponent": "alpha",
-  "Primer::PopoverComponent": "beta",
+  "Primer::PopoverComponent": "deprecated",
   "Primer::ProgressBarComponent": "deprecated",
   "Primer::SpinnerComponent": "beta",
   "Primer::StateComponent": "beta",
